services/ingest/requirements.txt
confluent-kafka==2.3.0
pymysql==1.1.0
psycopg2-binary==2.9.9
pymongo==4.6.0
